refactor(shared-data,api): module defs in module
This moves the python bindings for module defs to
opentrons_shared_data.module instead of opentrons_shared_data.modules to
match the name of the directory where the defs live
