+++ shared-data/python/opentrons_shared_data/module/__init__.py
@@ -1,4 +1,4 @@
-""" opentrons_shared_data.modules: functions and types for module defs """
+""" opentrons_shared_data.module: functions and types for module defs """
 
 import json
 from pathlib import Path
